adding fluids helper functions/utilities
